check if test_balsam already moved, more verbose tests, error checking

+++ conda/install-balsam.py
@@ -11,6 +11,7 @@ if int(sys.version[2]) >= 6:
     os.chdir('../balsam')
     subprocess.check_call('pip install -e .'.split())
     os.chdir('../libensemble')
-    os.rename('./conda/test_balsam.py',
-              './libensemble/tests/regression_tests/test_balsam.py')
+    if not os.path.isfile('./libensemble/tests/regression_tests/test_balsam.py'):
+        os.rename('./conda/test_balsam.py',
+                  './libensemble/tests/regression_tests/test_balsam.py')
     os.system('./conda/configure-balsam-test.sh')
